requirements.txt
Django==4.2
sqlparse==0.2.4
pymongo==3.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
pandas==2.1.3
openpyxl==3.1.2
PyPDF2==3.0.0
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
pytesseract==0.3.10
